* Wolk with external data provider

* Expanded data handler api; Fixed building tests

* Split handling of external device data and statuses

* Don't run tests when building

* Don't build and install gateway app

* Fixed routing of device messages

* Don't build and install app if used as subproject

* Newest WolkSDK-Cpp - fixing sensors & alarm messages with multiple references (#49)

* Change ExternalDataService to `make_shared`

* Fix setupGatewayDataService to not pass ref to invalid pointer

* Link the data service to subscribe to topics for config/actuation for gateway

* Added the `FileListener` class to `FileDownloadService`. (#52)

* Introduced FileListener

* Pass listener to service, and fix passed path

* Enabled Firmware Update

* FileListener control of files

* Add possiblity to add file to repo

* Removed fileCreationLambda

* Custom protocol and persistence (#53)

* Custom protocols; Publish external message from custom protocol as is

* FileSystem persistence

* sdk; Fixes includes and formatting

* Improved use of filesystem persistence;
Circular persistence

* Fixed publishing

* Persistence mechanism uses ordering

* Check messages retreived from buffer

* code review fixes

* Use recursive mutex for persistence (#56)

* Report update success & Store version in folder (#57)

* Report successful update on return true

* Save version file in separate folder

* Bugfix/blocking manager start (#59)

* New WolkSDK-Cpp with non-blocking log manager start

* Point to master

* Fix/json protocol sensor topics (#60)

* New SDK with fixed topics

* Fixed sensor topics

* Point to master commit of WolkSDK

* Fixed readings path when checking file size; (#61)

Fixed building tests

* Disable gw update for custom protocol (#62)

* Updated WolkSDK-Cpp submodule (#63)

* Bugfix/wolk destructor hanging (#64)

* New WolkSDK with Buffer::stop() method

* Call the stop method on buffer, and check if thread is joinable first

* New WolkSDK with AWS args based on Debug/Release (#65)

* Improvement/WolkSDK-Cpp with new setters and getters. (#66)

* Updated WolkSDK-Cpp submodule

* New WolkSDK with more getters

* Updated WolkSDK (reformat)

* Point to master WolkSDK-Cpp


* Improvement - Use FileSystem as source of truth about files (#67)

* Started implementing the FSFileRepository

* Completely remove the .hash-file

* Implemented the full unit tests for FSFileRepository

* Use FSFileRepository and not SQLiteFileRepository

* Applied the request code review changes

* Change to newest WolkSDK-Cpp (optional description field in payloads)

* Wolk with external data provider

* Expanded data handler api; Fixed building tests

* Split handling of external device data and statuses

* Don't run tests when building

* Don't build and install gateway app

* Fixed routing of device messages

* Don't build and install app if used as subproject

* Change ExternalDataService to `make_shared`

* Fix setupGatewayDataService to not pass ref to invalid pointer

* Link the data service to subscribe to topics for config/actuation for gateway

* Added the `FileListener` class to `FileDownloadService`. (#52)

* Introduced FileListener

* Pass listener to service, and fix passed path

* Enabled Firmware Update

* FileListener control of files

* Add possiblity to add file to repo

* Removed fileCreationLambda

* Custom protocol and persistence (#53)

* Custom protocols; Publish external message from custom protocol as is

* FileSystem persistence

* sdk; Fixes includes and formatting

* Improved use of filesystem persistence;
Circular persistence

* Fixed publishing

* Persistence mechanism uses ordering

* Check messages retreived from buffer

* code review fixes

* Use recursive mutex for persistence (#56)

* Report update success & Store version in folder (#57)

* Report successful update on return true

* Save version file in separate folder

* Fix/json protocol sensor topics (#60)

* New SDK with fixed topics

* Fixed sensor topics

* Point to master commit of WolkSDK

* Fixed readings path when checking file size; (#61)

Fixed building tests

* Disable gw update for custom protocol (#62)

* Improvement - Use FileSystem as source of truth about files (#67)

* Started implementing the FSFileRepository

* Completely remove the .hash-file

* Implemented the full unit tests for FSFileRepository

* Use FSFileRepository and not SQLiteFileRepository

* Applied the request code review changes

* Fixed enough after rebase to get it working

* Changed call of `addSensorReading` in tests to reflect new protocol changes

* Initialize the FSFileRepository later

* Fixed the initialization sequence for FSFileRepository

* Cleaned-up duplicates and made `build()` and `setupWith(External/Internal)Data()` method cleaner

* Create the device repositories only in internal

* Bump the version to v4.3.0

* Added the platform connection status getter and callback

* Update to newest WolkSDK-Cpp

* Added `POCO_INSTALL_DIR` value and reformated CMakeLists.txt

* Added the include dir too

* Fix files test

* Added licence headers; Formatting

+++ src/Wolk.cpp
@@ -16,31 +16,31 @@
 
 #include "Wolk.h"
 
-#include "connectivity/ConnectivityService.h"
-#include "model/ConfigurationSetCommand.h"
-#include "persistence/Persistence.h"
-#include "protocol/DataProtocol.h"
+#include "core/connectivity/ConnectivityService.h"
+#include "core/model/ConfigurationSetCommand.h"
+#include "core/persistence/Persistence.h"
+#include "core/protocol/DataProtocol.h"
+#include "core/protocol/RegistrationProtocol.h"
+#include "core/protocol/StatusProtocol.h"
+#include "core/protocol/json/JsonDFUProtocol.h"
+#include "core/protocol/json/JsonDownloadProtocol.h"
+#include "core/utilities/Logger.h"
 #include "protocol/GatewayDataProtocol.h"
 #include "protocol/GatewayFirmwareUpdateProtocol.h"
 #include "protocol/GatewayStatusProtocol.h"
 #include "protocol/GatewaySubdeviceRegistrationProtocol.h"
-#include "protocol/RegistrationProtocol.h"
-#include "protocol/StatusProtocol.h"
-#include "protocol/json/JsonDFUProtocol.h"
-#include "protocol/json/JsonDownloadProtocol.h"
 #include "repository/DeviceRepository.h"
 #include "repository/ExistingDevicesRepository.h"
 #include "repository/FileRepository.h"
-#include "service/DataService.h"
-#include "service/DeviceStatusService.h"
 #include "service/FileDownloadService.h"
 #include "service/FirmwareUpdateService.h"
-#include "service/GatewayDataService.h"
 #include "service/GatewayUpdateService.h"
 #include "service/KeepAliveService.h"
 #include "service/PublishingService.h"
 #include "service/SubdeviceRegistrationService.h"
-#include "utilities/Logger.h"
+#include "service/data/DataService.h"
+#include "service/data/GatewayDataService.h"
+#include "service/status/DeviceStatusService.h"
 
 #include <memory>
 #include <sstream>
@@ -62,16 +62,14 @@ WolkBuilder Wolk::newBuilder(GatewayDevice device)
     return WolkBuilder(std::move(device));
 }
 
-void Wolk::connect()
+bool Wolk::isConnectedToPlatform()
 {
-    connectToPlatform(true);
-    connectToDevices(true);
+    return m_connected;
 }
 
-void Wolk::disconnect()
+void Wolk::setPlatformConnectionStatusListener(const std::function<void(bool)>& platformConnectionStatusListener)
 {
-    addToCommandBuffer([=]() -> void { m_platformConnectivityService->disconnect(); });
-    addToCommandBuffer([=]() -> void { m_deviceConnectivityService->disconnect(); });
+    this->m_platformConnectionStatusListener = platformConnectionStatusListener;
 }
 
 void Wolk::addSensorReading(const std::string& reference, const std::string& value, unsigned long long rtc)
@@ -283,14 +281,6 @@ void Wolk::platformDisconnected()
     });
 }
 
-void Wolk::devicesDisconnected()
-{
-    addToCommandBuffer([=] {
-        notifyDevicesDisonnected();
-        connectToDevices(true);
-    });
-}
-
 void Wolk::gatewayUpdated()
 {
     addToCommandBuffer([=] {
@@ -307,19 +297,6 @@ void Wolk::gatewayUpdated()
             m_subdeviceRegistrationService->updatePostponedDevices();
         }
     });
-}
-
-void Wolk::deviceRegistered(const std::string& deviceKey)
-{
-    addToCommandBuffer([=] {
-        m_deviceStatusService->sendLastKnownStatusForDevice(deviceKey);
-        m_existingDevicesRepository->addDeviceKey(deviceKey);
-    });
-}
-
-void Wolk::deviceUpdated(const std::string& deviceKey)
-{
-    addToCommandBuffer([=] { m_deviceStatusService->sendLastKnownStatusForDevice(deviceKey); });
 }
 
 void Wolk::publishEverything()
@@ -365,13 +342,20 @@ void Wolk::updateGatewayAndDeleteDevices()
         //        m_gatewayUpdateService->updateGateway(m_device);
         shouldUpdate = false;
 
-        m_subdeviceRegistrationService->deleteDevicesOtherThan(m_existingDevicesRepository->getDeviceKeys());
+        if (m_subdeviceRegistrationService && m_device.getSubdeviceManagement().value() == SubdeviceManagement::GATEWAY)
+        {
+            m_subdeviceRegistrationService->deleteDevicesOtherThan(m_existingDevicesRepository->getDeviceKeys());
+        }
     }
 }
 
 void Wolk::notifyPlatformConnected()
 {
     LOG(INFO) << "Connection to platform established";
+
+    m_connected = true;
+    if (m_platformConnectionStatusListener)
+        m_platformConnectionStatusListener(m_connected);
 
     m_platformPublisher->connected();
 
@@ -385,30 +369,16 @@ void Wolk::notifyPlatformDisonnected()
 {
     LOG(INFO) << "Connection to platform lost";
 
+    m_connected = false;
+    if (m_platformConnectionStatusListener)
+        m_platformConnectionStatusListener(m_connected);
+
     m_platformPublisher->disconnected();
 
     if (m_keepAliveService)
     {
         m_keepAliveService->disconnected();
     }
-}
-
-void Wolk::notifyDevicesConnected()
-{
-    LOG(INFO) << "Connection to local bus established";
-
-    m_devicePublisher->connected();
-
-    m_deviceStatusService->connected();
-}
-
-void Wolk::notifyDevicesDisonnected()
-{
-    LOG(INFO) << "Connection to local bus lost";
-
-    m_devicePublisher->disconnected();
-
-    m_deviceStatusService->disconnected();
 }
 
 void Wolk::connectToPlatform(bool firstTime)
@@ -436,27 +406,6 @@ void Wolk::connectToPlatform(bool firstTime)
 
             std::this_thread::sleep_for(std::chrono::milliseconds(RECONNECT_DELAY_MSEC));
             connectToPlatform(false);
-        }
-    });
-}
-
-void Wolk::connectToDevices(bool firstTime)
-{
-    addToCommandBuffer([=]() -> void {
-        if (firstTime)
-            LOG(INFO) << "Connecting to local bus...";
-
-        if (m_deviceConnectivityService->connect())
-        {
-            notifyDevicesConnected();
-        }
-        else
-        {
-            if (firstTime)
-                LOG(INFO) << "Failed to connect to local bus";
-
-            std::this_thread::sleep_for(std::chrono::milliseconds(RECONNECT_DELAY_MSEC));
-            connectToDevices(false);
         }
     });
 }
